refactor: move effects to custom hooks
